Update Amount.amount from float to BigDecimal

+++ src/main/java/seedu/billboard/ui/charts/ExpenseTimelineChart.java
@@ -1,5 +1,6 @@
 package seedu.billboard.ui.charts;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
 import java.util.EnumMap;
@@ -25,7 +26,7 @@ public class ExpenseTimelineChart extends ExpenseChart<ExpenseTimeline> {
     private static final String FXML = "ExpenseTimelineChart.fxml";
 
     @FXML
-    private LineChart<String, Float> timelineChart;
+    private LineChart<String, BigDecimal> timelineChart;
 
     @FXML
     private CategoryAxis xAxis;
@@ -34,7 +35,7 @@ public class ExpenseTimelineChart extends ExpenseChart<ExpenseTimeline> {
     private NumberAxis yAxis;
 
     private final EnumMap<DateInterval, DateTimeFormatter> dateIntervalFormats;
-    private XYChart.Series<String, Float> series;
+    private XYChart.Series<String, BigDecimal> series;
 
 
     public ExpenseTimelineChart(ExpenseTimeline expenseTimeline) {
@@ -55,19 +56,19 @@ public class ExpenseTimelineChart extends ExpenseChart<ExpenseTimeline> {
     @Override
     public void onDataChange(ExpenseTimeline newData) {
         Map<DateRange, Amount> timeline = newData.getTimeline();
-        List<XYChart.Data<String, Float>> data = getData(timeline, newData.getDateInterval());
+        List<XYChart.Data<String, BigDecimal>> data = getData(timeline, newData.getDateInterval());
         series.getData().setAll(data);
     }
 
-    private List<XYChart.Data<String, Float>> getData(Map<DateRange, Amount> timeline, DateInterval interval) {
+    private List<XYChart.Data<String, BigDecimal>> getData(Map<DateRange, Amount> timeline, DateInterval interval) {
         return timeline.entrySet()
                 .stream()
                 .map(entry -> entryToData(entry, interval))
                 .collect(Collectors.toList());
     }
 
-    private XYChart.Data<String, Float> entryToData(Map.Entry<DateRange, Amount> entry,
-                                                               DateInterval interval) {
+    private XYChart.Data<String, BigDecimal> entryToData(Map.Entry<DateRange, Amount> entry,
+                                                         DateInterval interval) {
 
         return new XYChart.Data<>(formatDate(entry.getKey().getStartDate(), interval), entry.getValue().amount);
     }
